extracts email verification files to its own recipe
