+++ src/main/java/io/supertokens/emailpassword/EmailPassword.java
@@ -16,53 +16,37 @@
 
 package io.supertokens.emailpassword;
 
-import com.google.gson.Gson;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import io.supertokens.Main;
-import io.supertokens.emailpassword.exceptions.EmailAlreadyVerifiedException;
-import io.supertokens.emailpassword.exceptions.EmailVerificationInvalidTokenException;
 import io.supertokens.emailpassword.exceptions.ResetPasswordInvalidTokenException;
 import io.supertokens.emailpassword.exceptions.WrongCredentialsException;
-import io.supertokens.pluginInterface.emailpassword.EmailVerificationTokenInfo;
 import io.supertokens.pluginInterface.emailpassword.PasswordResetTokenInfo;
 import io.supertokens.pluginInterface.emailpassword.UserInfo;
-import io.supertokens.pluginInterface.emailpassword.exceptions.*;
+import io.supertokens.pluginInterface.emailpassword.exceptions.DuplicateEmailException;
+import io.supertokens.pluginInterface.emailpassword.exceptions.DuplicatePasswordResetTokenException;
+import io.supertokens.pluginInterface.emailpassword.exceptions.DuplicateUserIdException;
+import io.supertokens.pluginInterface.emailpassword.exceptions.UnknownUserIdException;
 import io.supertokens.pluginInterface.emailpassword.sqlStorage.EmailPasswordSQLStorage;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
 import io.supertokens.storageLayer.StorageLayer;
 import io.supertokens.utils.Utils;
-import io.supertokens.webserver.WebserverAPI;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import javax.servlet.ServletException;
 import java.security.NoSuchAlgorithmException;
 import java.security.SecureRandom;
 import java.security.spec.InvalidKeySpecException;
-import java.util.Base64;
 
 public class EmailPassword {
 
     public static final long PASSWORD_RESET_TOKEN_LIFETIME_MS =
             3600 * 1000; // this is related to the interval for the cronjob: DeleteExpiredPasswordResetTokens
 
-    public static final long EMAIL_VERIFICATION_TOKEN_LIFETIME_MS =
-            24 * 3600 * 1000; // this is related to the interval for the cronjob: DeleteExpiredEmailVerificationTokens
-
     private static long getPasswordResetTokenLifetime(Main main) {
         if (Main.isTesting) {
             return EmailPasswordTest.getInstance(main).getPasswordResetTokenLifetime();
         }
         return PASSWORD_RESET_TOKEN_LIFETIME_MS;
-    }
-
-    private static long getEmailVerificationTokenLifetime(Main main) {
-        if (Main.isTesting) {
-            return EmailPasswordTest.getInstance(main).getEmailVerificationTokenLifetime();
-        }
-        return EMAIL_VERIFICATION_TOKEN_LIFETIME_MS;
     }
 
     public static User signUp(Main main, @Nonnull String email, @Nonnull String password) throws
@@ -217,129 +201,16 @@ public class EmailPassword {
         return new User(info.id, info.email, info.timeJoined);
     }
 
-    public static String generateEmailVerificationToken(Main main, String userId)
-            throws InvalidKeySpecException, NoSuchAlgorithmException, StorageQueryException,
-            UnknownUserIdException, EmailAlreadyVerifiedException {
-
-        UserInfo user = StorageLayer.getEmailPasswordStorage(main).getUserInfoUsingId(userId);
-
-        if (user == null) {
-            throw new UnknownUserIdException();
-        }
-
-        if (user.isEmailVerified) {
-            throw new EmailAlreadyVerifiedException();
-        }
-
-        while (true) {
-
-            // we first generate a email verification token
-            byte[] random = new byte[64];
-            byte[] salt = new byte[64];
-
-            new SecureRandom().nextBytes(random);
-            new SecureRandom().nextBytes(salt);
-
-            int iterations = 1000;
-            String token = Utils
-                    .toHex(Utils.pbkdf2(Utils.bytesToString(random).toCharArray(), salt, iterations, 64 * 6));
-
-            // we make it URL safe:
-            token = Utils.convertToBase64(token);
-            token = token.replace("=", "");
-            token = token.replace("/", "");
-            token = token.replace("+", "");
-
-            String hashedToken = Utils.hashSHA256(token);
-
-            try {
-                StorageLayer.getEmailPasswordStorage(main).addEmailVerificationToken(
-                        new EmailVerificationTokenInfo(userId, hashedToken,
-                                System.currentTimeMillis() + getEmailVerificationTokenLifetime(main), user.email));
-                return token;
-            } catch (DuplicateEmailVerificationTokenException ignored) {
-            }
-        }
-    }
-
-    public static User verifyEmail(Main main, String token)
-            throws StorageQueryException, EmailVerificationInvalidTokenException, NoSuchAlgorithmException,
-            StorageTransactionLogicException {
-
-        String hashedToken = Utils.hashSHA256(token);
-
-        EmailPasswordSQLStorage storage = StorageLayer.getEmailPasswordStorage(main);
-
-        final EmailVerificationTokenInfo tokenInfo = storage.getEmailVerificationTokenInfo(hashedToken);
-        if (tokenInfo == null) {
-            throw new EmailVerificationInvalidTokenException();
-        }
-
-        final String userId = tokenInfo.userId;
-
-        try {
-            return storage.startTransaction(con -> {
-
-                EmailVerificationTokenInfo[] allTokens = storage
-                        .getAllEmailVerificationTokenInfoForUser_Transaction(con, userId);
-
-                EmailVerificationTokenInfo matchedToken = null;
-                for (EmailVerificationTokenInfo tok : allTokens) {
-                    if (tok.token.equals(hashedToken)) {
-                        matchedToken = tok;
-                        break;
-                    }
-                }
-
-                if (matchedToken == null) {
-                    throw new StorageTransactionLogicException(new EmailVerificationInvalidTokenException());
-                }
-
-                storage.deleteAllEmailVerificationTokensForUser_Transaction(con, userId);
-
-                if (matchedToken.tokenExpiry < System.currentTimeMillis()) {
-                    storage.commitTransaction(con);
-                    throw new StorageTransactionLogicException(new EmailVerificationInvalidTokenException());
-                }
-
-                UserInfo userInfo = storage.getUserInfoUsingId_Transaction(con, userId);
-
-                if (!userInfo.email.equals(tokenInfo.email)) {
-                    throw new StorageTransactionLogicException(new EmailVerificationInvalidTokenException());
-                }
-
-                storage.updateUsersIsEmailVerified_Transaction(con, userId, true);
-
-                storage.commitTransaction(con);
-                return new User(userInfo.id, userInfo.email, userInfo.timeJoined);
-            });
-        } catch (StorageTransactionLogicException e) {
-            if (e.actualException instanceof EmailVerificationInvalidTokenException) {
-                throw (EmailVerificationInvalidTokenException) e.actualException;
-            }
-            throw e;
-        }
-    }
-
-    public static boolean isEmailVerified(Main main, String userId) throws UnknownUserIdException,
-            StorageQueryException {
-        UserInfo user = StorageLayer.getEmailPasswordStorage(main).getUserInfoUsingId(userId);
-
-        if (user == null) {
-            throw new UnknownUserIdException();
-        }
-
-        return user.isEmailVerified;
-    }
-
     public static UserPaginationContainer getUsers(Main main, @Nullable String paginationToken, Integer limit,
-                                                   String timeJoinedOrder) throws StorageQueryException, IllegalArgumentException {
+                                                   String timeJoinedOrder)
+            throws StorageQueryException, IllegalArgumentException {
         UserInfo[] users;
         if (paginationToken == null) {
             users = StorageLayer.getEmailPasswordStorage(main).getUsers(limit + 1, timeJoinedOrder);
         } else {
             UserPaginationToken tokenInfo = UserPaginationToken.extractTokenInfo(paginationToken);
-            users = StorageLayer.getEmailPasswordStorage(main).getUsers(tokenInfo.userId, tokenInfo.timeJoined, limit + 1, timeJoinedOrder);
+            users = StorageLayer.getEmailPasswordStorage(main)
+                    .getUsers(tokenInfo.userId, tokenInfo.timeJoined, limit + 1, timeJoinedOrder);
         }
         String nextPaginationToken = null;
         int maxLoop = users.length;
