+++ src/main/java/io/supertokens/webserver/api/emailverification/VerifyEmailAPI.java
@@ -1,5 +1,5 @@
 /*
- *    Copyright (c) 2020, VRAI Labs and/or its affiliates. All rights reserved.
+ *    Copyright (c) 2021, VRAI Labs and/or its affiliates. All rights reserved.
  *
  *    This software is licensed under the Apache License, Version 2.0 (the
  *    "License") as published by the Apache Software Foundation.
@@ -14,15 +14,15 @@
  *    under the License.
  */
 
-package io.supertokens.webserver.api.emailpassword;
+package io.supertokens.webserver.api.emailverification;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import io.supertokens.Main;
-import io.supertokens.emailpassword.EmailPassword;
 import io.supertokens.emailpassword.User;
 import io.supertokens.emailpassword.exceptions.EmailVerificationInvalidTokenException;
+import io.supertokens.emailverification.EmailVerification;
 import io.supertokens.pluginInterface.emailpassword.exceptions.UnknownUserIdException;
 import io.supertokens.pluginInterface.exceptions.StorageQueryException;
 import io.supertokens.pluginInterface.exceptions.StorageTransactionLogicException;
@@ -63,7 +63,7 @@ public class VerifyEmailAPI extends WebserverAPI {
         }
 
         try {
-            User user = EmailPassword.verifyEmail(super.main, token);
+            User user = EmailVerification.verifyEmail(super.main, token);
 
             JsonObject result = new JsonObject();
             result.addProperty("status", "OK");
@@ -85,7 +85,7 @@ public class VerifyEmailAPI extends WebserverAPI {
         String userId = InputParser.getQueryParamOrThrowError(req, "userId", false);
 
         try {
-            boolean isVerified = EmailPassword.isEmailVerified(super.main, userId);
+            boolean isVerified = EmailVerification.isEmailVerified(super.main, userId);
 
             JsonObject result = new JsonObject();
             result.addProperty("status", "OK");
