+++ src/test/java/io/supertokens/test/DeleteExpiredEmailVerificationTokensCronjobTest.java
@@ -21,6 +21,7 @@ import io.supertokens.cronjobs.CronTaskTest;
 import io.supertokens.cronjobs.deleteExpiredEmailVerificationTokens.DeleteExpiredEmailVerificationTokens;
 import io.supertokens.emailpassword.EmailPassword;
 import io.supertokens.emailpassword.User;
+import io.supertokens.emailverification.EmailVerification;
 import io.supertokens.pluginInterface.STORAGE_TYPE;
 import io.supertokens.pluginInterface.emailpassword.EmailVerificationTokenInfo;
 import io.supertokens.storageLayer.StorageLayer;
@@ -63,14 +64,14 @@ public class DeleteExpiredEmailVerificationTokensCronjobTest {
 
         User user = EmailPassword.signUp(process.getProcess(), "test1@example.com", "password");
 
-        String tok = EmailPassword.generateEmailVerificationToken(process.getProcess(), user.id);
-        String tok2 = EmailPassword.generateEmailVerificationToken(process.getProcess(), user.id);
+        String tok = EmailVerification.generateEmailVerificationToken(process.getProcess(), user.id);
+        String tok2 = EmailVerification.generateEmailVerificationToken(process.getProcess(), user.id);
 
-        io.supertokens.emailpassword.EmailPasswordTest.getInstance(process.getProcess())
+        io.supertokens.emailverification.EmailVerificationTest.getInstance(process.getProcess())
                 .setEmailVerificationTokenLifetime(10);
 
-        EmailPassword.generateEmailVerificationToken(process.getProcess(), user.id);
-        EmailPassword.generateEmailVerificationToken(process.getProcess(), user.id);
+        EmailVerification.generateEmailVerificationToken(process.getProcess(), user.id);
+        EmailVerification.generateEmailVerificationToken(process.getProcess(), user.id);
 
         assert (StorageLayer.getEmailPasswordStorage(process.getProcess())
                 .getAllEmailVerificationTokenInfoForUser(user.id).length == 4);
